fix: make the overlay doesn't block pointer event
Previously, even when the navbar is closed (the overlay doesn't appear)
we can't select anything inside the page because I didn't put hidden
(display none) in the overlay.

If I put the hidden class in the overlay and toggle that class with js,
the opacity animation would not applied. So I thought to just not put
the hidden class.

But turns out the overlay blocked our interaction to the web pages even
when the navbar menu is closed.

So now, I decided to just set the `pointer-events` to `none` if the nav
menu is closed to let us click and interact anything in the page.